Move utilities for releasing main thread to the shared package
