+++ .changeset/stupid-buttons-relax.md
@@ -1,0 +1,5 @@
+---
+"@itwin/presentation-shared": patch
+---
+
+Added utility functions for releasing main thread: `releaseMainThread` and `createMainThreadReleaseOnTimePassedHandler`.
